+++ packages/runtime/test/src/interpreter/opcode-list.ts
@@ -169,16 +169,17 @@ import {
 import {
 	ALGORAND_ACCOUNT_MIN_BALANCE,
 	ASSET_CREATION_FEE,
-	CurveTypeEnum,
 	blockFieldTypes,
+	CurveTypeEnum,
 	DEFAULT_STACK_ELEM,
 	MAX_UINT8,
 	MAX_UINT64,
 	MaxTEALVersion,
 	MIN_UINT8,
+	seedLength,
+	TxFieldEnum,
 	vrfVerifyFieldTypes,
 	ZERO_ADDRESS,
-	seedLength,
 } from "../../../src/lib/constants";
 import {
 	bigEndianBytesToBigInt,
@@ -2102,15 +2103,6 @@ describe("Teal Opcodes", function () {
 				assert.deepEqual(TXN_OBJ.rekey, stack.pop());
 			});
 
-			it("should throw error on FirstValidTime", function () {
-				execExpectError(
-					stack,
-					[],
-					new Txn(["FirstValidTime"], 1, interpreter),
-					RUNTIME_ERRORS.TEAL.REJECTED_BY_LOGIC
-				);
-			});
-
 			it("should push txn NumAppArgs to stack", function () {
 				const op = new Txn(["NumAppArgs"], 1, interpreter);
 				op.execute(stack);
@@ -2555,6 +2547,58 @@ describe("Teal Opcodes", function () {
 			});
 		});
 
+		describe("FirstValidTime", function () {
+			this.beforeEach(function () {
+				interpreter = new Interpreter();
+				interpreter.runtime = new Runtime([]);
+				interpreter.runtime.ctx.tx = cloneDeep(TXN_OBJ);
+				interpreter.tealVersion = MaxTEALVersion;
+			})
+
+			it("Should throw exception if txn.lv - txn.fv > 1000 + 1", function () {
+				interpreter.runtime.ctx.tx.fv = 1999; //last valid = 3000, 3000 - (1999 - 1) > 1001
+				const op = new Txn([TxFieldEnum.FirstValidTime], 1, interpreter);
+				assert.throws(() => op.execute(stack));
+			});
+
+			it("Should push txn correct timestamp to stack", function () {
+				interpreter.runtime.ctx.tx.lv = TXN_OBJ.fv + 500;
+				const op = new Txn([TxFieldEnum.FirstValidTime], 1, interpreter);
+				const expectedResult = interpreter.runtime.getBlock(TXN_OBJ.fv - 1).timestamp;
+				op.execute(stack);
+				assert.equal(expectedResult, stack.pop());
+			});
+
+			it("Should throw exception if teal version < 7", function () {
+				interpreter.tealVersion = 6;
+				expectRuntimeError(() =>
+				 new Txn([TxFieldEnum.FirstValidTime], 1, interpreter),
+				 RUNTIME_ERRORS.TEAL.UNKNOWN_TRANSACTION_FIELD);
+			});
+
+			it("Should throw exception if round number is negative", function () {
+				interpreter.runtime.ctx.tx.fv = 0;
+				const op = new Txn([TxFieldEnum.FirstValidTime], 1, interpreter);
+				assert.throws(() => op.execute(stack));
+			});
+
+			it("Should throw exception if round number is 0 even if last valid is low", function () {
+				interpreter.runtime.ctx.tx.fv = 1;
+				interpreter.runtime.ctx.tx.lv = 100;
+				const op = new Txn([TxFieldEnum.FirstValidTime], 1, interpreter);
+				assert.throws(() => op.execute(stack));
+			});
+
+			it("Should not throw error on FirstValidTime", function () {
+				execExpectError(
+					stack,
+					[],
+					new Txn([TxFieldEnum.FirstValidTime], 1, interpreter),
+					RUNTIME_ERRORS.TEAL.REJECTED_BY_LOGIC
+				);
+			});
+		});
+
 		describe("Gtxn", function () {
 			before(function () {
 				const tx = interpreter.runtime.ctx.tx;
@@ -2563,6 +2607,14 @@ describe("Teal Opcodes", function () {
 				// https://developer.algorand.org/docs/reference/transactions/
 				const tx2 = { ...tx, fee: 2222, apas: [3033, 4044], apfa: [5005, 6006, 7077] };
 				interpreter.runtime.ctx.gtxs = [tx, tx2];
+			});
+
+			it("Should push to the stack correct block timestamp", function(){
+				const op = new Gtxn(["0",TxFieldEnum.FirstValidTime], 1, interpreter);
+				const firstTxFirstValid = interpreter.runtime.ctx.gtxs[0].fv?? 0;
+				const expextedResult = interpreter.runtime.getBlock(firstTxFirstValid - 1).timestamp;
+				op.execute(stack);
+				assert.equal(expextedResult, stack.pop());
 			});
 
 			it("push fee from 2nd transaction in group", function () {
@@ -5070,8 +5122,8 @@ describe("Teal Opcodes", function () {
 				assert.deepEqual(
 					stack.pop(),
 					new Uint8Array([
-						3, 35, 69, 118, 1, 35, 69, 117, 3, 35, 69, 118, 1, 35, 69, 117, 2, 70, 138, 236, 2,
-						70, 138, 236, 2, 70, 138, 236, 2, 70, 138, 236,
+						3, 35, 69, 118, 1, 35, 69, 117, 3, 35, 69, 118, 1, 35, 69, 117, 2, 70, 138, 236,
+						2, 70, 138, 236, 2, 70, 138, 236, 2, 70, 138, 236,
 					])
 				);
 
@@ -5171,8 +5223,8 @@ describe("Teal Opcodes", function () {
 				assert.deepEqual(
 					stack.pop(),
 					new Uint8Array([
-						1, 35, 69, 118, 0, 35, 69, 118, 255, 35, 69, 118, 1, 35, 69, 119, 0, 0, 0, 0, 0, 0,
-						0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+						1, 35, 69, 118, 0, 35, 69, 118, 255, 35, 69, 118, 1, 35,
+						69, 119, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 					])
 				);
 			});
@@ -5244,8 +5296,8 @@ describe("Teal Opcodes", function () {
 					new Uint8Array([
 						151, 200, 103, 239, 94, 230, 133, 186, 92, 4, 163, 133, 89, 34, 193, 80, 86, 64,
 						223, 27, 83, 94, 252, 230, 80, 125, 26, 177, 77, 155, 56, 124, 72, 239, 162, 240,
-						235, 209, 133, 37, 238, 179, 103, 90, 241, 149, 73, 143, 244, 119, 43, 196, 247, 89,
-						13, 249, 250, 58, 240, 46, 253, 28, 210, 100,
+						235, 209, 133, 37, 238, 179, 103, 90, 241, 149, 73, 143, 244, 119, 43, 196, 247,
+						89, 13, 249, 250, 58, 240, 46, 253, 28, 210, 100,
 					])
 				);
 			});
@@ -5645,8 +5697,8 @@ describe("Teal Opcodes", function () {
 				assert.deepEqual(
 					stack.pop(),
 					new Uint8Array([
-						227, 23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 22, 19,
-						23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 28,
+						227, 23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 22,
+						19, 23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 22, 19, 23, 18, 28,
 					])
 				);
 			});
@@ -5686,8 +5738,8 @@ describe("Teal Opcodes", function () {
 				assert.deepEqual(
 					stack.pop(),
 					new Uint8Array([
-						94, 220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137, 254,
-						220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137,
+						94, 220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137,
+						254, 220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137, 254, 220, 186, 137,
 					])
 				);
 			});
@@ -6294,8 +6346,10 @@ describe("Teal Opcodes", function () {
 			interpreter = new Interpreter();
 			interpreter.runtime = new Runtime([]);
 			interpreter.runtime.ctx.tx = TXN_OBJ;
-			interpreter.tealVersion = MaxTEALVersion; // set tealversion to latest (to support all tx fields)
-			// a) 'apas' represents 'foreignAssets', b) 'apfa' represents 'foreignApps' (id's of foreign apps)
+			interpreter.tealVersion = MaxTEALVersion;
+			// set tealversion to latest (to support all tx fields)
+			// a) 'apas' represents 'foreignAssets',
+			// b) 'apfa' represents 'foreignApps' (id's of foreign apps)
 			// https://developer.algorand.org/docs/reference/transactions/
 			const tx2 = { ...TXN_OBJ, fee: 2222, apas: [3033, 4044], apfa: [5005, 6006, 7077] };
 			interpreter.runtime.ctx.gtxs = [TXN_OBJ, tx2];
@@ -7598,7 +7652,7 @@ describe("Teal Opcodes", function () {
 			assert.notEqual(seedRound1, seedRound2);
 		});
 
-		it("Should fail when accesing block 0 even if last valid is low", function(){
+		it("Should fail when accesing block 0 even if last valid is low", function () {
 			interpreter.runtime.ctx.tx.lv = 100;
 			interpreter.runtime.ctx.tx.fv = 5;
 			stack.push(0n);
@@ -7606,7 +7660,7 @@ describe("Teal Opcodes", function () {
 			assert.throws(() => op.execute(stack));
 		});
 
-		it("Should return seed for a block that is within boundries and exist", function(){
+		it("Should return seed for a block that is within boundries and exist", function () {
 			stack.push(FIRST_VALID - 1n);
 			const op = new Block([blockFieldTypes.BlkSeed], 1, interpreter);
 			op.execute(stack);
@@ -7614,7 +7668,7 @@ describe("Teal Opcodes", function () {
 			assert.equal((result as Buffer).length, seedLength);
 		});
 
-		it("Should return correct cost", function(){
+		it("Should return correct cost", function () {
 			stack.push(FIRST_VALID - 1n);
 			const op = new Block([blockFieldTypes.BlkSeed], 1, interpreter);
 			const cost = op.execute(stack);
@@ -7651,7 +7705,7 @@ describe("Teal Opcodes", function () {
 			assert.deepEqual(1n, stack.pop());
 			assert.deepEqual(expectedResult, stack.pop());
 		});
-		it("Should throw error if pubKey size not equal 32", function(){
+		it("Should throw error if pubKey size not equal 32", function () {
 			const op = new VrfVerify([vrfVerifyFieldTypes.VrfAlgorand], 1);
 			const wrongSizePubKey = Buffer.from("b6b4699f87d56126c9117");
 			stack.push(wrongSizePubKey);
@@ -7659,7 +7713,7 @@ describe("Teal Opcodes", function () {
 			stack.push(message);
 			expectRuntimeError(() => op.execute(stack), RUNTIME_ERRORS.TEAL.INVALID_PUB_KEY_LENGTH);
 		})
-		it("Should throw error if proof size not equal 80", function(){
+		it("Should throw error if proof size not equal 80", function () {
 			const op = new VrfVerify([vrfVerifyFieldTypes.VrfAlgorand], 1);
 			const wrongSizeProof = Buffer.from("b6b4699f87d56126c9117");
 			stack.push(publicKey);
@@ -7667,7 +7721,7 @@ describe("Teal Opcodes", function () {
 			stack.push(message);
 			expectRuntimeError(() => op.execute(stack), RUNTIME_ERRORS.TEAL.INVALID_PROOF_LENGTH);
 		})
-		it("Should return correct cost", function(){
+		it("Should return correct cost", function () {
 			const op = new VrfVerify([vrfVerifyFieldTypes.VrfAlgorand], 1);
 			stack.push(publicKey);
 			stack.push(proof);
@@ -7675,7 +7729,7 @@ describe("Teal Opcodes", function () {
 			const cost = op.execute(stack);
 			assert.equal(cost, 5700);
 		})
-		it("Should throw error when field = VrfStandard", function(){
+		it("Should throw error when field = VrfStandard", function () {
 			const op = new VrfVerify([vrfVerifyFieldTypes.VrfStandard], 1);
 			stack.push(publicKey);
 			stack.push(proof);